requirements.txt
asgiref==3.9.1
Django==5.2.5
django-mathfilters==1.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
